# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Game On Tabletop API
  description: REST API for Game On Tabletop — a platform for organising tabletop
    game nights, tracking game collections, scheduling events, and managing group
    RSVPs.
  version: 1.0.0
paths:
...
  /api/series/{series_id}/polls:
    post:
      tags:
      - polls
      summary: Create an availability poll for a series (organiser only)
      operationId: create_poll_api_series__series_id__polls_post
      security:
      - HTTPBearer: []
      parameters:
      - name: series_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Series Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PollCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollOut'
        '401':
          description: Unauthorized
        '403':
          description: Not an organiser
        '404':
          description: Series not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - polls
      summary: List polls for a series
      operationId: list_polls_api_series__series_id__polls_get
      security:
      - HTTPBearer: []
      parameters:
      - name: series_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Series Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PollOut'
                title: Response List Polls Api Series  Series Id  Polls Get
        '401':
          description: Unauthorized
        '403':
          description: Not a member
        '404':
          description: Series not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/polls/{poll_id}:
    get:
      tags:
      - polls
      summary: Get a poll with options and response counts
      operationId: get_poll_api_polls__poll_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: poll_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Poll Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollOut'
        '401':
          description: Unauthorized
        '403':
          description: Not a member
        '404':
          description: Poll not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/polls/{poll_id}/options:
    post:
      tags:
      - polls
      summary: Add a date option to a poll (organiser only)
      operationId: add_poll_option_api_polls__poll_id__options_post
      security:
      - HTTPBearer: []
      parameters:
      - name: poll_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Poll Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PollOptionCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollOptionOut'
        '400':
          description: Poll is not open
        '401':
          description: Unauthorized
        '403':
          description: Not an organiser
        '404':
          description: Poll not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/polls/{poll_id}/options/{option_id}:
    delete:
      tags:
      - polls
      summary: Remove a date option from a poll (organiser only)
      operationId: remove_poll_option_api_polls__poll_id__options__option_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: poll_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Poll Id
      - name: option_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Option Id
      responses:
        '204':
          description: Successful Response
        '400':
          description: Poll is not open
        '401':
          description: Unauthorized
        '403':
          description: Not an organiser
        '404':
          description: Poll or option not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/polls/{poll_id}/respond:
    post:
      tags:
      - polls
      summary: Submit or update availability responses for a poll
      operationId: respond_to_poll_api_polls__poll_id__respond_post
      security:
      - HTTPBearer: []
      parameters:
      - name: poll_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Poll Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PollRespond'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollOut'
        '400':
          description: Poll is not open
        '401':
          description: Unauthorized
        '403':
          description: Not a member
        '404':
          description: Poll not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/polls/{poll_id}/resolve:
    post:
      tags:
      - polls
      summary: Resolve a poll by picking the winning date and creating an occurrence
        (organiser only)
      operationId: resolve_poll_api_polls__poll_id__resolve_post
      security:
      - HTTPBearer: []
      parameters:
      - name: poll_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Poll Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PollResolve'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollOut'
        '400':
          description: Poll already resolved or invalid option
        '401':
          description: Unauthorized
        '403':
          description: Not an organiser
        '404':
          description: Poll not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/polls/{poll_id}/guest-link:
    post:
      tags:
      - polls
      summary: Create a guest link for unauthenticated poll responses (organiser only)
      operationId: create_poll_guest_link_api_polls__poll_id__guest_link_post
      security:
      - HTTPBearer: []
      parameters:
      - name: poll_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Poll Id
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestTokenOut'
        '401':
          description: Unauthorized
        '403':
          description: Not an organiser
        '404':
          description: Poll not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/occurrences/{occurrence_id}/guest-link:
    post:
      tags:
      - polls
      summary: Create a guest RSVP link for an occurrence (organiser only)
      operationId: create_occurrence_guest_link_api_occurrences__occurrence_id__guest_link_post
      security:
      - HTTPBearer: []
      parameters:
      - name: occurrence_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Occurrence Id
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestTokenOut'
        '401':
          description: Unauthorized
        '403':
          description: Not an organiser
        '404':
          description: Occurrence not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    GuestTokenOut:
      properties:
        token:
          type: string
          title: Token
        url:
          type: string
          title: Url
        expires_at:
          type: string
          format: date-time
          title: Expires At
      type: object
      required:
      - token
      - url
      - expires_at
      title: GuestTokenOut
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OptionResponse:
      properties:
        option_id:
          type: string
          format: uuid
          title: Option Id
        response:
          type: string
          enum:
          - 'yes'
          - 'no'
          - maybe
          title: Response
      type: object
      required:
      - option_id
      - response
      title: OptionResponse
    PollCreate:
      properties:
        title:
          type: string
          maxLength: 200
          minLength: 1
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        deadline:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deadline
        options:
          items:
            $ref: '#/components/schemas/PollOptionCreate'
          type: array
          title: Options
      type: object
      required:
      - title
      title: PollCreate
    PollOptionCreate:
      properties:
        proposed_date:
          type: string
          format: date
          title: Proposed Date
        start_time:
          type: string
          format: time
          title: Start Time
        end_time:
          anyOf:
          - type: string
            format: time
          - type: 'null'
          title: End Time
        location_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Location Id
        display_order:
          type: integer
          minimum: 0.0
          title: Display Order
          default: 0
      type: object
      required:
      - proposed_date
      - start_time
      title: PollOptionCreate
    PollOptionOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        poll_id:
          type: string
          format: uuid
          title: Poll Id
        proposed_date:
          type: string
          format: date
          title: Proposed Date
        start_time:
          type: string
          format: time
          title: Start Time
        end_time:
          anyOf:
          - type: string
            format: time
          - type: 'null'
          title: End Time
        location_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Location Id
        display_order:
          type: integer
          title: Display Order
        response_counts:
          additionalProperties:
            type: integer
          type: object
          title: Response Counts
      type: object
      required:
      - id
      - poll_id
      - proposed_date
      - start_time
      - end_time
      - location_id
      - display_order
      title: PollOptionOut
    PollOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        series_id:
          type: string
          format: uuid
          title: Series Id
        title:
          type: string
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        deadline:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deadline
        status:
          type: string
          title: Status
        chosen_option_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Chosen Option Id
        created_by:
          type: string
          format: uuid
          title: Created By
        created_at:
          type: string
          format: date-time
          title: Created At
        options:
          items:
            $ref: '#/components/schemas/PollOptionOut'
          type: array
          title: Options
        my_responses:
          additionalProperties:
            type: string
          type: object
          title: My Responses
      type: object
      required:
      - id
      - series_id
      - title
      - description
      - deadline
      - status
      - chosen_option_id
      - created_by
      - created_at
      title: PollOut
    PollResolve:
      properties:
        chosen_option_id:
          type: string
          format: uuid
          title: Chosen Option Id
      type: object
      required:
      - chosen_option_id
      title: PollResolve
    PollRespond:
      properties:
        responses:
          items:
            $ref: '#/components/schemas/OptionResponse'
          type: array
          title: Responses
      type: object
      required:
      - responses
      title: PollRespond
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer